cleanup of namespaces
cleans up namespaces and unused using statements
