+++ src/Ucommerce.Sitecore.Install.Steps/AddHeadlessToIgnoreUrlPrefixes.cs
@@ -4,10 +4,9 @@ using System.Threading.Tasks;
 using System.Xml;
 using System.Xml.XPath;
 using Ucommerce.Installer;
-using Ucommerce.Sitecore.Install;
-using Ucommerce.Sitecore.Installer.FileExtensions;
+using Ucommerce.Sitecore.Install.Steps.FileExtensions;
 
-namespace Ucommerce.Sitecore.Installer.Steps
+namespace Ucommerce.Sitecore.Install.Steps
 {
     public class AddHeadlessToIgnoreUrlPrefixes : IStep
     {
