+++ src/Ucommerce.Sitecore.Install.Steps/ComposeMoveSitecoreConfigIncludes.cs
@@ -1,10 +1,9 @@
 ﻿using System;
 using System.IO;
 using Ucommerce.Installer;
-using Ucommerce.Sitecore.Install;
-using Ucommerce.Sitecore.Installer.FileExtensions;
+using Ucommerce.Sitecore.Install.Steps.FileExtensions;
 
-namespace Ucommerce.Sitecore.Installer.Steps
+namespace Ucommerce.Sitecore.Install.Steps
 {
     public class ComposeMoveSitecoreConfigIncludes : AggregateStep
     {
