+++ src/Ucommerce.Sitecore.Install.Steps/EnableSitecoreCompatibilityApp.cs
@@ -2,9 +2,8 @@ using System;
 using System.IO;
 using System.Threading.Tasks;
 using Ucommerce.Installer;
-using Ucommerce.Sitecore.Install;
 
-namespace Ucommerce.Sitecore.Installer.Steps
+namespace Ucommerce.Sitecore.Install.Steps
 {
     public class EnableSitecoreCompatibilityApp : IStep
     {
