+++ src/Ucommerce.Sitecore.Install.Steps/InitializeObjectFactory.cs
@@ -1,8 +1,7 @@
 ﻿using System.Threading.Tasks;
 using Ucommerce.Installer;
-using Ucommerce.Sitecore.Install;
 
-namespace Ucommerce.Sitecore.Installer.Steps
+namespace Ucommerce.Sitecore.Install.Steps
 {
     public class InitializeObjectFactory : IStep
     {
