+++ src/Ucommerce.Sitecore.Install.Steps/InstallDatabaseUcommerce.cs
@@ -1,10 +1,9 @@
 ﻿using System.IO;
 using System.Threading.Tasks;
 using Ucommerce.Installer;
-using Ucommerce.Sitecore.Install;
-using Ucommerce.Sitecore.Installer.FileExtensions;
+using Ucommerce.Sitecore.Install.Steps.FileExtensions;
 
-namespace Ucommerce.Sitecore.Installer.Steps
+namespace Ucommerce.Sitecore.Install.Steps
 {
     public class InstallDatabaseUcommerce : IStep
     {
