+++ src/Ucommerce.Sitecore.Install.Steps/SitecoreInstallationConnectionStringLocator.cs
@@ -1,0 +1,24 @@
+﻿using System.Configuration;
+using Ucommerce.Installer;
+
+namespace Ucommerce.Sitecore.Install.Steps
+{
+    public class SitecoreInstallationConnectionStringLocator : InstallationConnectionStringLocator
+    {
+        private readonly string _connectionString;
+
+        public SitecoreInstallationConnectionStringLocator(string connectionString)
+        {
+            _connectionString = connectionString;
+        }
+
+        public override string LocateConnectionString()
+        {
+            if (string.IsNullOrEmpty(_connectionString))
+                throw new ConfigurationErrorsException(
+                    "No connection string was given");
+
+            return _connectionString;
+        }
+    }
+}
